+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,4 +1,5 @@
-import { Component, OnInit, Output, EventEmitter } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
+import { FormControl, FormGroup } from '@angular/forms';
 
 import { ProductModel } from '../../models';
 import { ProductsService } from '../../services';
@@ -6,18 +7,32 @@ import { CartService } from '../../../cart/services';
 
 @Component({
     selector: 'app-product-list',
-    templateUrl: './product-list.component.html'
+    templateUrl: './product-list.component.html',
+    styleUrls: ['./product-list.component.scss']
 })
 export class ProductListComponent implements OnInit {
+    orders = [
+        { value: 'name', name: 'Product name' },
+        { value: 'price', name: 'Product price' },
+        { value: 'quantity', name: 'Quantity' },
+    ];
 
-    products: ProductModel[];
+    form = new FormGroup({
+        order: new FormControl(this.orders[0]),
+    });
+
+    products$: Promise<ProductModel[]>;
+
+    get sortOrder(): string {
+        return this.form.value.order.value;
+    }
 
     constructor(
         private productsService: ProductsService,
         private cartService: CartService) { }
 
     ngOnInit() {
-        this.products = this.productsService.getProducts();
+        this.products$ = this.productsService.getProducts();
     }
 
     onBuy(product: ProductModel) {