+++ src/app/products/enums/product-category.emun.ts
@@ -1,10 +1,4 @@
 export enum ProductCategory {
-    Accessories,
-    Automotive,
-    Baby,
-    Beauty,
-    Books,
-    Clothes,
-    Shoes,
-    Sport
+    Magazines = 'Magazines',
+    Books = 'Books'
 }