+++ src/app/products/services/products.service.ts
@@ -5,30 +5,34 @@ import { ProductCategory } from '../enums';
 
 
 @Injectable({
-  providedIn: 'root'
+    providedIn: 'root'
 })
 export class ProductsService {
 
-  constructor() { }
-
-  getProducts(): ProductModel[] {
-    return [
-      new ProductModel(
-        '1', 'A Brief History of Time',
-        13.59, ProductCategory.Books,
-        1,
-        'A landmark volume in science writing by one of the great minds of our time, Stephen Hawking’s book...'),
-      new ProductModel(
-        '2', 'A Heartbreaking Work of Staggering Genius',
-        10.84, ProductCategory.Books,
-        2,
-        'A book that redefines both family and narrative for the twenty-first century. A Heartbreaking Work of Staggering Genius is...'),
-      new ProductModel(
-        '3',
-        'Alice`s Adventures in Wonderland & Through the Looking-Glass',
-        5.95, ProductCategory.Books,
-        3,
-        'In 1862 Charles Lutwidge Dodgson, a shy Oxford mathematician with a stammer, created a story about a little girl tumbling...'),
+    products: ProductModel[] = [
+        new ProductModel(
+            '1', 'A Brief History of Time',
+            13.59, ProductCategory.Books,
+            12,
+            'A landmark volume in science writing by one of the great minds of our time, Stephen Hawking’s book...'),
+        new ProductModel(
+            '2', 'A Heartbreaking Work of Staggering Genius',
+            10.84, ProductCategory.Books,
+            20,
+            'A book that redefines both family and narrative for the twenty-first century. A Heartbreaking Work of Staggerin...'),
+        new ProductModel(
+            '3',
+            'Alice`s Adventures in Wonderland & Through the Looking-Glass',
+            5.95, ProductCategory.Books,
+            3,
+            'In 1862 Charles Lutwidge Dodgson, a shy Oxford mathematician with a stammer, created a story about a little girl tumbling...'),
     ];
-  }
+
+    constructor() { }
+
+    getProducts(): Promise<ProductModel[]> {
+        return new Promise<ProductModel[]>((resolve, reject) => {
+            resolve(this.products);
+        });
+    }
 }