+++ src/app/shared/shared.module.ts
@@ -2,14 +2,17 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { HighlightDirective, HostClickDirective } from './directives';
+import { OrderByPipe } from './pipes/order-by.pipe';
 @NgModule({
   declarations: [
     HighlightDirective,
-    HostClickDirective
+    HostClickDirective,
+    OrderByPipe
   ],
   exports: [
     HighlightDirective,
-    HostClickDirective
+    HostClickDirective,
+    OrderByPipe
   ],
   imports: [
     CommonModule