+++ changelog.md
@@ -1,4 +1,18 @@
-<a name="1.3.0"></a>
+<a name="1.4.0"></a>
+
+## 1.4.0 (2019-02-07)
+
+### Added
+ SharedModule
+ - OrderByPipe
+ 
+### Notes
+Pipes:
+- async pipe is used in the ProductListComponent
+- orderBy pipe is used in the ProductListComponent
+- uppercase pipes are used in the ProductComponent
+- titlecase pipe is used in the ProductComponent
+- currency pipes are used in the ProductComponent, CartItemComponent, CartComponent
 
 ## 1.3.0 (2019-02-04)